core/serializers/user.py
from rest_framework import serializers

from core.utils.user import User

class UserSerializer(serializers.ModelSerializer):
    class Meta:
        model = User
        fields = ['id', 'username', 'email', 'is_staff']
