requirements.txt
channels[daphne]
coverage
django
django-admin-extra-buttons
django_cleanup
django-cors-headers
django-filter
django-storages
django-unfold
djangorestframework
djangorestframework-simplejwt
drf-redesign
drf-spectacular
gunicorn
loguru
markdown
pillow
psycopg2-binary
pydantic
pylint
pylint-django
pytest
pytest-django
python-dotenv
whitenoise
